Feat: fixed the review comments

+++ src/Apha.FPSApps/Apha.FPSApps.Web/Areas/FPS/Models/CompareStaff2Item.cs
@@ -29,7 +29,7 @@ namespace Apha.FPSApps.Web.Areas.FPS.Models
         public double? Time { get; set; }
 
         [GridColumn(Width = 90, Type = GridColumnType.GbpValue)]
-        public double? Cost { get; set; }
+        public decimal? Cost { get; set; }
 
         /// <summary>Composite key used by the grid delete action: workgroup|jobcode|project|month|staffid</summary>
         [GridColumn(IsVisible = false)]
